pyproject.toml

[project]
name = "aussiebb_outage_watcher"
version = "0.2.0"
authors = [{ name = "James Hodgkinson", email = "james@terminaloutcomes.com" }]
description = ""
requires-python = "<4.0,>=3.10"
dependencies = [
    "pydantic>=2.11.5",
 "schedule>=1.2.2",
 "pyaussiebb==0.1.6",
]

[project.scripts]
"aussiebb-outage-watcher" = "aussiebb_outage_watcher.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 200

[dependency-groups]
dev = [
    "ruff>=0.11.12",
    "mypy>=1.16.0",
    "pytest>=8.3.5",
    "pylint-pydantic>=0.3.5",
]
